[Add] #20 - 프로필 뷰 추가
