+++ DaangnMarket/DaangnMarket/Presentation/MyDaangn/ViewController/ProfileViewController.swift
@@ -2,7 +2,7 @@
 //  ProfileViewController.swift
 //  DaangnMarket
 //
-//  Created by 김민 on 2023/04/27.
+//  Created by 김민 on 2023/04/28.
 //
 
 import UIKit
@@ -12,14 +12,12 @@ import Then
 
 final class ProfileViewController: UIViewController {
 
-    // MARK: - View Life Cycle    
+    // MARK: - View Life Cycle
     
     override func viewDidLoad() {
         super.viewDidLoad()
 
         setBackgroundColor()
-        setLayout()
-        setNavigationBar()
     }
 }
 
@@ -28,26 +26,6 @@ extension ProfileViewController {
     // MARK: - Custom Methods
     
     private func setBackgroundColor() {
-        view.backgroundColor = .white
-    }
-    
-    private func setLayout() {
-        
-    }
-    
-    private func setNavigationBar() {
-        navigationItem.title = "프로필"
-        navigationItem.leftBarButtonItem = UIBarButtonItem(image: UIImage(systemName: "chevron.left"), style: .plain, target: self, action: #selector(backButtonDidTap))
-        navigationItem.rightBarButtonItem = UIBarButtonItem(image: UIImage(systemName: "square.and.arrow.up"), style: .plain, target: self, action: nil)
-        
-        navigationController?.navigationBar.tintColor = .black
-        
-        setNavigationBackground()
-    }
-    
-    // MARK: - @objc Methods
-    
-    @objc private func backButtonDidTap() {
-        navigationController?.popViewController(animated: true)
+        view.backgroundColor = .yellow
     }
 }
